removing bw/login.php
git-svn-id: file:///home/jaliste/hcvol/hcvol/trunk@5468 4f49b16c-3c57-40ef-8e22-7452baa6c33a

+++ htdocs/bw/addcomments.php
@@ -49,7 +49,7 @@ if (!CheckStatus("Active")) { // only Active member can send a Message
 	 exit (0);
 }
 
-$IdMember = GetParam("cid", 0);
+$IdMember = (int)GetParam("cid", 0);
 if ($IdMember==$_SESSION['IdMember']) {
 	$errcode = "ErrorNoCommentOnYourSelf";
 	DisplayError(ww($errcode, $IdMember));
